fix(nra): the battery's exact lint, and two fixture assertions that were claims about the source
`scripts/check-clippy-complete.sh`: **903 of 903 workspace targets across 27 of
27 crates, all 903 compiled by this run, 0 diagnostics.**

That wrapper rather than the bare form, and the reason is this run. The bare
`cargo clippy --workspace --all-targets --all-features -- -D warnings` came back
green in **0.70 s** having printed `Checking axeyum-solver` -- which is the exact
shape CLAUDE.md warns about, a gate passing over code it never compiled. The
wrapper says how many targets it examined, and it had to touch all 14,392 build
inputs first because the target dir carried no freshness manifest. The green
above is the wrapper's, not the bare form's.

Five real findings, all this lane's:

* `CadPolicy` now carries four bools, so `struct_excessive_bools` fires. Allowed
  with a reason rather than restructured: this is a POLICY TABLE, and the lint's
  usual fix -- collapse the bools into an enum of states -- would destroy exactly
  the property every arm is built on, that two arms differ in EXACTLY one field.
* `the_single_cell_arm_differs_in_exactly_the_route` passed 100 lines once it
  read the whole arm table; allowed with a reason.
* `assert!(!CAD_DEFAULT.algebraic_witness, ...)` is const-folded, and clippy is
  right to reject it: a const assertion is a claim about a literal, and it would
  not fail if the default were repointed at some other arm. Replaced by a
  RUNTIME read of the shipped arm out of the table, which is the stronger
  statement and the form the neighbouring assertions already use.
* two redundant `as usize` casts in the new fuzz seed class.
* two fixture assertions in `nra_algebraic_witness.rs` were `assert!` over
  LITERALS (`25 * 2 > 49`, `1393² < 2·985²`) -- compile-time claims about the
  source text, not tests. Both now read the SAME locals the query is built from
  (`coef`/`limit`, `pn`/`pd`), so the check follows the fixture instead of
  restating it: change the query's numbers and the arithmetic check moves with
  them. That is a better test than the one clippy rejected, not a workaround for
  the lint.

Also green in this batch: `cargo fmt --all --check` 0; default-features
`cargo check --workspace --all-targets` 0 (56 s of real checking, every crate);
`config_registry` tests 18 passed 0 failed; `check-config-registry-staleness.py`
0 unexplained (33 stale rows, all 33 accepted with written reasons).

Refs ADR-2134.

Agent: nra-algebraic-witness

+++ crates/axeyum-solver/tests/nra_differential_fuzz.rs
@@ -400,7 +400,7 @@ impl Instance {
         let two_vars = rng.below(2) == 0;
         let num_vars = if two_vars { 2 } else { 1 };
         let last = num_vars - 1;
-        let c = NON_SQUARES[(rng.below(NON_SQUARES.len() as u64)) as usize];
+        let c = NON_SQUARES[rng.below(NON_SQUARES.len() as u64)];
 
         let mut atoms: Vec<Atom> = Vec::new();
 
@@ -457,7 +457,7 @@ impl Instance {
                     Monomial {
                         num: i128::from(q),
                         den: 1,
-                        factors: vec![last; degree as usize],
+                        factors: vec![last; degree],
                     },
                     Monomial {
                         num: i128::from(p),
